# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones frecuentes de docentes y estudiantes en EduGo",
        "title": "EduGo API Mobile",
        "termsOfService": "http://edugo.com/terms/",
        "contact": {
            "name": "Equipo EduGo",
            "email": "soporte@edugo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene lista de materiales filtrados por unidad y materia con progreso del estudiante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener lista de materiales",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de unidad académica",
                        "name": "unit_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID de materia",
                        "name": "subject_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "all",
                            "new",
                            "in_progress",
                            "completed"
                        ],
                        "type": "string",
                        "description": "Estado del material",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo material educativo y retorna URL firmada para subir PDF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Crear nuevo material",
                "parameters": [
                    {
                        "description": "Datos del material",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/CreateMaterialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene información completa de un material incluyendo URL firmada para descargar PDF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener detalle de material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/assessment": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene las preguntas del quiz SIN respuestas correctas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener cuestionario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AssessmentResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/assessment/attempts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Valida respuestas del quiz y registra puntaje del estudiante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Registrar intento de evaluación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Respuestas del estudiante",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RecordAttemptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AttemptResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/progress": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra el progreso del estudiante en la lectura de un material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Actualizar progreso de lectura",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Progreso del estudiante",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateProgressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene progreso y puntajes de todos los estudiantes para un material (solo docentes)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener estadísticas del material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialStatsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No eres docente de este material",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene el resumen generado por IA desde MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener resumen del material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialSummaryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Resumen no disponible aún",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateMaterialRequest": {
            "type": "object",
            "required": [
                "description",
                "file_name",
                "file_size",
                "subject_id",
                "title",
                "unit_id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Material base sobre programación en Pascal"
                },
                "file_name": {
                    "type": "string",
                    "example": "introduccion-pascal.pdf"
                },
                "file_size": {
                    "type": "integer",
                    "example": 2048576
                },
                "subject_id": {
                    "type": "string",
                    "example": "subject-uuid-123"
                },
                "title": {
                    "type": "string",
                    "example": "Introducción a Pascal"
                },
                "unit_id": {
                    "type": "string",
                    "example": "unit-uuid-456"
                }
            }
        },
        "CreateMaterialResponse": {
            "type": "object",
            "properties": {
                "material_id": {
                    "type": "string",
                    "example": "uuid-123"
                },
                "max_file_size_bytes": {
                    "type": "integer",
                    "example": 104857600
                },
                "status": {
                    "type": "string",
                    "example": "created"
                },
                "upload_url": {
                    "type": "string",
                    "example": "https://s3.amazonaws.com/edugo-materials-prod/upload-url-mock"
                },
                "upload_url_expires_at": {
                    "type": "string",
                    "example": "2025-01-29T12:15:00Z"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Field 'title' is required"
                }
            }
        },
...
        "MaterialListResponse": {
            "type": "object",
            "properties": {
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MaterialSummary"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "RecordAttemptRequest": {
            "type": "object",
            "required": [
                "answers"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Answer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token con formato: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}