# internal/handlers/materials.go
package handlers

import (
	"net/http"
	"time"

	_ "github.com/edugo/api-mobile/internal/models/request" // Usado en comentarios de Swagger
	"github.com/edugo/api-mobile/internal/models/response"
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

// GetMaterials godoc
// @Summary Obtener lista de materiales
// @Description Obtiene lista de materiales filtrados por unidad y materia con progreso del estudiante
// @Tags Materials
// @Accept json
// @Produce json
// @Param unit_id query string false "ID de unidad académica"
// @Param subject_id query string false "ID de materia"
// @Param status query string false "Estado del material" Enums(all, new, in_progress, completed)
// @Security BearerAuth
// @Success 200 {object} response.MaterialListResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /materials [get]
func GetMaterials(c *gin.Context) {
	// Obtener parámetros de query
	_ = c.Query("unit_id")    // unitID - usado para filtrar en producción
	_ = c.Query("subject_id") // subjectID - usado para filtrar en producción
	status := c.DefaultQuery("status", "all")

	// TODO: Implementar lógica real con PostgreSQL
	// Por ahora, retornar datos mock

	mockMaterials := []response.MaterialSummary{
		{
			ID:          "material-uuid-1",
			Title:       "Introducción a Pascal",
			SubjectName: "Programación",
			UnitName:    "5.º A - Programación",
			Status:      "new",
			Progress:    0,
			HasSummary:  true,
			HasQuiz:     true,
			PublishedAt: time.Date(2025, 1, 15, 12, 0, 0, 0, time.UTC),
		},
		{
			ID:          "material-uuid-2",
			Title:       "Estructuras de Control",
			SubjectName: "Programación",
			UnitName:    "5.º A - Programación",
			Status:      "in_progress",
			Progress:    45,
			HasSummary:  true,
			HasQuiz:     false,
			PublishedAt: time.Date(2025, 1, 20, 10, 30, 0, 0, time.UTC),
		},
	}

	// Filtrar por status (simulado)
	filteredMaterials := mockMaterials
	if status != "all" {
		filtered := []response.MaterialSummary{}
		for _, m := range mockMaterials {
			if m.Status == status {
				filtered = append(filtered, m)
			}
		}
		filteredMaterials = filtered
	}

	c.JSON(http.StatusOK, response.MaterialListResponse{
		Materials: filteredMaterials,
		Total:     len(filteredMaterials),
		Page:      1,
	})
}

// GetMaterialDetail godoc
// @Summary Obtener detalle de material
// @Description Obtiene información completa de un material incluyendo URL firmada para descargar PDF
// @Tags Materials
// @Accept json
// @Produce json
// @Param id path string true "ID del material"
// @Security BearerAuth
// @Success 200 {object} response.MaterialDetailResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Router /materials/{id} [get]
func GetMaterialDetail(c *gin.Context) {
	materialID := c.Param("id")

	// TODO: Implementar lógica real
	// - Consultar PostgreSQL para metadatos
	// - Generar URL firmada de S3
	// - Verificar existencia de resumen en material_summary_link

	// Mock response
	mockDetail := response.MaterialDetailResponse{
		Material: response.MaterialDetail{
			ID:          materialID,
			Title:       "Introducción a Pascal",
			Description: "Material base sobre historia y sintaxis del lenguaje Pascal para 5.º año",
			AuthorName:  "Prof. García",
			SubjectName: "Programación",
			FileSize:    2048576,
			MyProgress:  0,
			CreatedAt:   time.Date(2025, 1, 15, 12, 0, 0, 0, time.UTC),
		},
		PDFURL:          "https://s3.amazonaws.com/edugo-materials-prod/presigned-url-mock?expires=...",
		PDFURLExpiresAt: time.Now().Add(15 * time.Minute),
		HasSummary:      true,
		HasQuiz:         true,
	}

	c.JSON(http.StatusOK, mockDetail)
}

// CreateMaterial godoc
// @Summary Crear nuevo material
// @Description Crea un nuevo material educativo y retorna URL firmada para subir PDF
// @Tags Materials
// @Accept json
// @Produce json
// @Param body body request.CreateMaterialRequest true "Datos del material"
// @Security BearerAuth
// @Success 201 {object} response.CreateMaterialResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Router /materials [post]
func CreateMaterial(c *gin.Context) {
	// TODO: Parsear request body
	// TODO: Validar permisos del docente
	// TODO: Crear material en PostgreSQL
	// TODO: Generar URL firmada de S3

	// Mock response
	materialID := uuid.New().String()
	uploadURL := "https://s3.amazonaws.com/edugo-materials-prod/upload-url-mock?X-Amz-Signature=..."

	c.JSON(http.StatusCreated, gin.H{
		"status":                "created",
		"material_id":           materialID,
		"upload_url":            uploadURL,
		"upload_url_expires_at": time.Now().Add(15 * time.Minute),
		"max_file_size_bytes":   104857600, // 100 MB
	})
}

// UploadComplete godoc
// @Summary Notificar que el upload se completó
// @Description Registra versión del material y publica evento para procesamiento
// @Tags Materials
// @Accept json
// @Produce json
// @Param id path string true "ID del material"
// @Param body body request.UploadCompleteRequest true "Datos del archivo subido"
// @Security BearerAuth
// @Success 202 {object} response.UploadCompleteResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Router /materials/{id}/upload-complete [post]
func UploadComplete(c *gin.Context) {
	materialID := c.Param("id")

	// TODO: Registrar versión en material_version
	// TODO: Calcular file_hash
	// TODO: Verificar deduplicación
	// TODO: Publicar evento material_uploaded a RabbitMQ

	c.JSON(http.StatusAccepted, gin.H{
		"status":                 "processing",
		"message":                "Material en procesamiento. Te notificaremos cuando el resumen esté listo.",
		"estimated_time_minutes": 3,
		"material": gin.H{
			"id":         materialID,
			"title":      "Introducción a Pascal",
			"status":     "processing",
			"created_at": time.Now(),
		},
	})
}

// GetMaterialSummary godoc
// @Summary Obtener resumen del material
// @Description Obtiene el resumen generado por IA desde MongoDB
// @Tags Materials
// @Accept json
// @Produce json
// @Param id path string true "ID del material"
// @Security BearerAuth
// @Success 200 {object} response.MaterialSummaryResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse "Resumen no disponible aún"
// @Router /materials/{id}/summary [get]
func GetMaterialSummary(c *gin.Context) {
	materialID := c.Param("id")

	// TODO: Consultar PostgreSQL para obtener mongo_document_id
	// TODO: Consultar MongoDB para obtener resumen completo

	// Mock response con estructura completa
	mockResponse := response.MaterialSummaryResponse{
		Sections: []response.SummarySection{
			{
				Title:                "Contexto Histórico",
				Content:              "Pascal fue desarrollado por Niklaus Wirth en 1970 como lenguaje de enseñanza. Su diseño enfatiza la programación estructurada y la claridad del código.",
				Difficulty:           "basic",
				EstimatedTimeMinutes: 5,
				Order:                1,
			},
			{
				Title:                "Características Principales",
				Content:              "Pascal introduce tipificación fuerte, procedimientos y funciones bien definidos, y estructuras de control claras. Es un lenguaje compilado que genera código eficiente.",
				Difficulty:           "medium",
				EstimatedTimeMinutes: 8,
				Order:                2,
			},
		},
		Glossary: []response.GlossaryTerm{
			{
				Term:       "Compilador",
				Definition: "Programa que traduce código fuente a código máquina ejecutable por la computadora.",
				Order:      1,
			},
			{
				Term:       "Tipificación Fuerte",
				Definition: "Sistema de tipos que previene operaciones entre tipos incompatibles, detectando errores en tiempo de compilación.",
				Order:      2,
			},
		},
		ReflectionQuestions: []string{
			"¿Qué ventajas aportó Pascal sobre lenguajes anteriores como FORTRAN o COBOL?",
			"¿Por qué es importante la tipificación fuerte en un lenguaje de programación?",
			"¿Cómo ayuda Pascal a aprender buenos hábitos de programación?",
		},
		ProcessingMetadata: response.ProcessingMetadata{
			NLPProvider:           "openai",
			Model:                 "gpt-4",
			TokensUsed:            3500,
			ProcessingTimeSeconds: 45,
			Language:              "es",
			PromptVersion:         "v1.2",
		},
	}

	_ = materialID // Suprimir warning
	c.JSON(http.StatusOK, mockResponse)
}

// GetAssessment godoc
// @Summary Obtener cuestionario
// @Description Obtiene las preguntas del quiz SIN respuestas correctas
// @Tags Materials
// @Accept json
// @Produce json
// @Param id path string true "ID del material"
// @Security BearerAuth
// @Success 200 {object} response.AssessmentResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Router /materials/{id}/assessment [get]
func GetAssessment(c *gin.Context) {
	materialID := c.Param("id")

	// TODO: Consultar PostgreSQL para obtener mongo_document_id
	// TODO: Consultar MongoDB con proyección que EXCLUYA correct_answer y feedback

	// Mock response (SIN correct_answer ni feedback) - estructura completa
	mockResponse := response.AssessmentResponse{
		Title:            "Cuestionario: Introducción a Pascal",
		Description:      "Evaluación de conceptos básicos sobre el lenguaje Pascal y sus características principales.",
		TotalQuestions:   5,
		TotalPoints:      100,
		PassingScore:     70,
		TimeLimitMinutes: 15,
		Questions: []response.Question{
			{
				ID:         "q1",
				Text:       "¿Qué es un compilador?",
				Type:       "multiple_choice",
				Difficulty: "basic",
				Points:     20,
				Order:      1,
				Options: []response.QuestionOption{
					{ID: "a", Text: "Un programa que traduce código fuente a código máquina"},
					{ID: "b", Text: "Un tipo de variable en Pascal"},
					{ID: "c", Text: "Un editor de texto especializado"},
					{ID: "d", Text: "Una estructura de control de flujo"},
				},
			},
			{
				ID:         "q2",
				Text:       "¿Cuál es una característica principal de Pascal?",
				Type:       "multiple_choice",
				Difficulty: "medium",
				Points:     20,
				Order:      2,
				Options: []response.QuestionOption{
					{ID: "a", Text: "Tipificación débil"},
					{ID: "b", Text: "Tipificación fuerte"},
					{ID: "c", Text: "No tiene tipos de datos"},
					{ID: "d", Text: "Solo soporta números enteros"},
				},
			},
			{
				ID:         "q3",
				Text:       "Pascal fue diseñado principalmente para:",
				Type:       "multiple_choice",
				Difficulty: "basic",
				Points:     20,
				Order:      3,
				Options: []response.QuestionOption{
					{ID: "a", Text: "Desarrollo de videojuegos"},
					{ID: "b", Text: "Enseñanza de programación estructurada"},
					{ID: "c", Text: "Programación de sistemas operativos"},
					{ID: "d", Text: "Desarrollo web"},
				},
			},
			{
				ID:         "q4",
				Text:       "¿Quién creó el lenguaje Pascal?",
				Type:       "multiple_choice",
				Difficulty: "basic",
				Points:     20,
				Order:      4,
				Options: []response.QuestionOption{
					{ID: "a", Text: "Dennis Ritchie"},
					{ID: "b", Text: "Niklaus Wirth"},
					{ID: "c", Text: "Bjarne Stroustrup"},
					{ID: "d", Text: "James Gosling"},
				},
			},
			{
				ID:         "q5",
				Text:       "La tipificación fuerte en Pascal ayuda a:",
				Type:       "multiple_choice",
				Difficulty: "medium",
				Points:     20,
				Order:      5,
				Options: []response.QuestionOption{
					{ID: "a", Text: "Ejecutar programas más rápido"},
					{ID: "b", Text: "Detectar errores en tiempo de compilación"},
					{ID: "c", Text: "Reducir el tamaño del código"},
					{ID: "d", Text: "Permitir cualquier tipo de operación"},
				},
			},
		},
	}

	_ = materialID // Suprimir warning
	c.JSON(http.StatusOK, mockResponse)
}

// RecordAttempt godoc
// @Summary Registrar intento de evaluación
// @Description Valida respuestas del quiz y registra puntaje del estudiante
// @Tags Materials
// @Accept json
// @Produce json
// @Param id path string true "ID del material"
// @Param body body request.RecordAttemptRequest true "Respuestas del estudiante"
// @Security BearerAuth
// @Success 200 {object} response.AttemptResultResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Router /materials/{id}/assessment/attempts [post]
func RecordAttempt(c *gin.Context) {
	// TODO: Parsear request body
	// TODO: Obtener preguntas CON respuestas correctas y feedback de MongoDB
	// TODO: Validar cada respuesta comparando con correct_answer
	// TODO: Generar DetailedFeedback usando feedback.correct o feedback.incorrect
	// TODO: Calcular puntaje sumando points de respuestas correctas
	// TODO: Persistir en quiz_attempt + quiz_attempt_answer
	// TODO: Publicar evento assessment_attempt_recorded a RabbitMQ

	// Mock response con feedback detallado personalizado
	mockResponse := response.AttemptResultResponse{
		Score:       80.0,
		TotalPoints: 100.0,
		Passed:      true,
		DetailedFeedback: []response.QuestionFeedback{
			{
				QuestionID:      "q1",
				IsCorrect:       true,
				YourAnswer:      "a",
				FeedbackMessage: "¡Correcto! Un compilador traduce código fuente a código máquina. Esto permite que el programa se ejecute de forma eficiente en el procesador.",
			},
			{
				QuestionID:      "q2",
				IsCorrect:       true,
				YourAnswer:      "b",
				FeedbackMessage: "¡Correcto! La tipificación fuerte es una de las características principales de Pascal, diseñada para detectar errores en tiempo de compilación.",
			},
			{
				QuestionID:      "q3",
				IsCorrect:       true,
				YourAnswer:      "b",
				FeedbackMessage: "¡Correcto! Pascal fue diseñado específicamente para enseñar programación estructurada de forma clara y didáctica.",
			},
			{
				QuestionID:      "q4",
				IsCorrect:       true,
				YourAnswer:      "b",
				FeedbackMessage: "¡Correcto! Niklaus Wirth creó Pascal en 1970, nombrándolo en honor al matemático Blaise Pascal.",
			},
			{
				QuestionID:      "q5",
				IsCorrect:       false,
				YourAnswer:      "a",
				FeedbackMessage: "Incorrecto. La tipificación fuerte ayuda principalmente a detectar errores en tiempo de compilación, no a ejecutar más rápido. Revisa el concepto de verificación de tipos.",
			},
		},
	}

	c.JSON(http.StatusOK, mockResponse)
}

// UpdateProgress godoc
// @Summary Actualizar progreso de lectura
// @Description Registra el progreso del estudiante en la lectura de un material
// @Tags Materials
// @Accept json
// @Produce json
// @Param id path string true "ID del material"
// @Param body body request.UpdateProgressRequest true "Progreso del estudiante"
// @Security BearerAuth
// @Success 200 {object} response.SuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Router /materials/{id}/progress [patch]
func UpdateProgress(c *gin.Context) {
	// TODO: Parsear request body
	// TODO: Upsert en reading_log con GREATEST para progress

	c.JSON(http.StatusOK, gin.H{
		"message": "Progreso actualizado exitosamente",
		"status":  "success",
	})
}

// GetMaterialStats godoc
// @Summary Obtener estadísticas del material
// @Description Obtiene progreso y puntajes de todos los estudiantes para un material (solo docentes)
// @Tags Materials
// @Accept json
// @Produce json
// @Param id path string true "ID del material"
// @Security BearerAuth
// @Success 200 {object} response.MaterialStatsResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse "No eres docente de este material"
// @Router /materials/{id}/stats [get]
func GetMaterialStats(c *gin.Context) {
	materialID := c.Param("id")

	// TODO: Validar que usuario es docente/autor
	// TODO: Query complejo PostgreSQL para obtener progreso + intentos

	// Mock response
	c.JSON(http.StatusOK, gin.H{
		"material": gin.H{
			"id":    materialID,
			"title": "Introducción a Pascal",
		},
		"summary": gin.H{
			"total_students":   25,
			"not_started":      5,
			"in_progress":      12,
			"completed":        8,
			"average_progress": 62.4,
			"average_score":    75.5,
			"completion_rate":  32.0,
		},
		"students": []gin.H{
			{
				"id":           "student-uuid-1",
				"name":         "Ana García",
				"progress":     100,
				"latest_score": 95,
				"last_access":  time.Now().Add(-24 * time.Hour),
				"status":       "completed",
			},
		},
	})
}